refactor(console): fix duplicated subscription requests
